Started using only CellIds to represent blocks. Ram usage down to 200MB when not generating the world. Have to lower peaks of ram usage due to the "meshed" 2darray when greedy-meshing. Bugs with the composing of the mesh, just some indices thing gone wrong

+++ src/voxelengine/block/CellId.java
@@ -2,11 +2,11 @@ package voxelengine.block;
 
 public class CellId {
     
-    public static int ID_AIR = Integer.MAX_VALUE;
-    public static int ID_GRASS = 0;
-    public static int ID_DIRT = 1;
-    public static int ID_WOOD = 2; 
-    public static int ID_LEAVES = 3;
-    public static int ID_STONE = 4;
+    public static byte ID_AIR = Byte.MAX_VALUE;
+    public static byte ID_GRASS = 0;
+    public static byte ID_DIRT = 1;
+    public static byte ID_WOOD = 2; 
+    public static byte ID_LEAVES = 3;
+    public static byte ID_STONE = 4;
     
 }
